Rename source file for better diff
