+++ packages/tracing/src/index.bundle.ts
@@ -58,7 +58,7 @@ import type { Integration } from '@sentry/types';
 import { GLOBAL_OBJ } from '@sentry/utils';
 
 import { BrowserTracing } from './browser';
-import { addExtensionMethods } from './hubextensions';
+import { addExtensionMethods } from './extensions';
 
 export { Span } from '@sentry/core';
